- Spiel mit Gui alles machen  - Single Player: Song still player much later (casue gameMode.startGame(this) is called too late but theres no real way to fix it without changing a bit ) - load and multi windows dont close alone,
Gui generel: Gui has mutliple windows when the game starts haha... (gefixt as we speak)
- machmal wenn undo , kommt eine fehler meldung von tui obwhol die gui es gemacht hat (throw exceptios is there but still idk weird shit , bild 1) beim zweites mal geht es ok , sonst ist undo redo ok

-   wenn spiel zuende ist, kommt dieser fehler, wahrlschanlich passiert das alles, weil die load und multi windows sich nicht schliessen...(bild 2)(as we speak i think the porblem is that i put too many interrupts everywhere, maybe its interrupting / starting / asking for input in an input sessions causing one input to be thrown which is "empty" since gui is being used,) , problem bei TUi lösen und nicht Gui anfassen (aber das mach schon ein undo redo nur halt null exceptions aber es geht)

- conclutions: New game bei single modus, rick roll erst später, bei multi window blebien offen, wahlschanlich das null excpetions passiert in bei modus
- bei load game:
- load geht klar, bei load passiert den selebn bug auch , ist nicht zu sagen was genau es passieren könnte,
bei load game ist dieser game over fehler nicht passiert, ist aber trozdem nicht zu sagen ob es nochmal passieren würde.

-----
Game spielen aber nur mit tui:
- starting new game in multi, spiel anfang ganz normal, keine Abnormalität, karten legen geht, ohne "ask for player name" in mitte von prozess
- draw geht auch
- undo redo gehen hier oben error
- game enden geht auch klar, nur gui will die anderen fenstern nicht schlieesen und es macht condation racing mit tui manchmal wegen dieser "htread" aber ohne die geht nichts , naja.
Bei fehler eingaben geht es auch ok, erlaubt spieler nochmal zu spielen bei syntax, bei karten platzt fehler ist skill issue, spieler wird dann geweschselt
all in all is Tui Solide. Alle Events erwähne/putzen aber, single modus geht auch

---
Mish mash spielen:

- fragen gehen ok? hab gemerkt dass es manchmal hängt, also heute morgen war es so, vielleicht geht es wieder keine ahnung
- wenn man ein fehler in der gui macht dann will die tui nicht mehr gehen (grund: loop wird gestürtzt und die while loop für player turn für tui wird dann immer false und nie wieder true , muss ein event wieder triggern oder ähnliches zu aktivieren. (tui printet weiter nimmet aber nicht mehr)

sonst geht
music nicht getestet
