+++ src/main/scala/Main.scala
@@ -1,5 +1,5 @@
 import aview.Tui
-import aview.gui.GameGuiTui
+import aview.gui.Gui
 import com.google.inject.Guice
 import controller.GameControllerInterface
 
@@ -7,7 +7,8 @@ object Main extends App {
     private val injector = Guice.createInjector(new JsonModule)
     val controller = injector.getInstance(classOf[GameControllerInterface])
     val tui = new Tui(controller)
-    val gui = new GameGuiTui(controller)
+    val gui = new Gui(controller)
 
     controller.startGame()
+   
 }
